requirement.txt
streamlit
google-genai
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
pypdf
faiss-cpu
sentence-transformers